requirements.txt
fastapi==0.110.0
motor==3.3.2
python-dotenv==1.0.1
redis==5.0.3
uvicorn==0.29.0
colorlog==6.8.2
PyYAML==6.0.1
aiohttp==3.9.3
python-dateutil==2.9.0.post0
pytz==2024.1
google-api-python-client==2.123.0